+++ CHANGELOG.md
@@ -8,6 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx&utm_medium=paddle-node-sdk) for information about changes to the Paddle Billing platform, the Paddle API, and other developer tools.
 
+## 0.6.0 - 2024-02-20
+
+> **Breaking changes:** This version includes major improvements that introduce breaking changes. These are called out below.
+
+### Removed
+
+- **Breaking change:** Removed `*Includes` entity in favour of standard entities.
+  
+  - Use `Price` instead of `PriceWithIncludes`
+  - Use `Product` instead of `ProductWithIncludes`
+  - Use `Subscription` instead of `SubscriptionIncludes`
+  - Use `Transaction` instead of `TransactionIncludes`
+
+---
+
 ## 0.5.0 - 2024-02-16
 
 > **Breaking changes:** This version includes major improvements that introduce breaking changes. These are called out below.
@@ -15,6 +30,8 @@ Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx
 ### Changed
 
 - We removed the shared entities between API and Notification as we foresee them diverging. No Action required for this change
+
+---
 
 ## 0.4.0 - 2024-02-14